validation.py moved to commands
create.py, update.py - validation moved to functions beginning
validation.py - removed transaction bit
