+++ link_shortener/commands/update.py
@@ -8,7 +8,7 @@ import os
 from sqlalchemy import and_
 
 from link_shortener.core.exceptions import NotFoundException
-from link_shortener.core.validation import endpoint_duplicity_check
+from link_shortener.commands.validation import endpoint_duplicity_check
 from link_shortener.models import links, salts
 
 
@@ -26,6 +26,8 @@ async def check_update_form(request, link_id):
 
 async def update_link(request, link_id, data):
     async with request.app.engine.acquire() as conn:
+        await endpoint_duplicity_check(conn, data)
+
         trans = await conn.begin()
         link_update = links.update().where(links.columns['id'] == link_id)
         query = await conn.execute(links.select().where(
@@ -35,8 +37,6 @@ async def update_link(request, link_id, data):
         if not link_data:
             await trans.close()
             raise NotFoundException
-
-        await endpoint_duplicity_check(conn, trans, data)
 
         if data['password']:
             salt = os.urandom(32)
